+++ frontend/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import TopBar from "./components/TopBar";
+import TopBar from "../components/TopBar";
 import AuthProvider from "./providers/AuthProvider";
 import MUIThemeProvider from "./providers/MUIThemeProvider";
 import { EmotionCacheProvider } from "./providers/EmotionCacheProvider";
